Workaround for StringFromFormat-under-linux problem, random cleanup.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@1491 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DolphinWX/Src/GameListCtrl.cpp
@@ -628,7 +628,7 @@ void CGameListCtrl::OnProperties(wxCommandEvent& WXUNUSED (event))
 
 void CGameListCtrl::MultiCompressCB(const char* text, float percent, void* arg)
 {
-    wxString textString(wxString::Format(wxT("%s (%i/%i) - %s"), m_currentFilename.c_str(), m_currentItem+1, m_numberItem, text));
+    wxString textString(wxString::Format(wxT("%s (%i/%i) - %s"), m_currentFilename.c_str(), (int)m_currentItem+1, (int)m_numberItem, text));
 
     percent = (((float)m_currentItem) + percent) / (float)m_numberItem;
     wxProgressDialog* pDialog = (wxProgressDialog*)arg;
